The health check does not require ASP.NET Core
Quartz.AspNetCore carries <FrameworkReference Include="Microsoft.AspNetCore.App" />
for the HTTP API, and a framework reference reaches the nuspec. The health check
was the only reason most workers referenced that package, so a worker on a
dotnet/runtime image took ASP.NET Core with it and failed to start.

QuartzHealthCheck reads IScheduler.Status and probes the job store; its only two
usings were Microsoft.Extensions.DependencyInjection and
Microsoft.Extensions.Diagnostics.HealthChecks, and core already had the first.
So the check, its target record, its options and the four registration methods
move to src/Quartz/Diagnostics/HealthChecks/, and Quartz takes
Microsoft.Extensions.Diagnostics.HealthChecks as a package reference. The
namespace stays Quartz and every method stays an extension method, so no call
site changes; the class hosting them is QuartzHealthCheckExtensions rather than
QuartzAspNetCoreConfigurationExtensions.

Quartz.AspNetCore keeps its framework reference for the HTTP API and the
dashboard, and keeps the half of the health-check story that really is ASP.NET
Core's: MapHealthChecks and the ResultStatusCodes mapping that decides what a
degraded scheduler answers over HTTP.

The two test fixtures move to Quartz.Tests.Unit, which references no ASP.NET
Core anything - that project compiling is the guard - beside a new
HealthCheckPackagingTest that fails if Quartz.csproj ever grows a
FrameworkReference again.

Closes #3532
