+++ docs/documentation/quartz-4.x/operations.md
@@ -652,10 +652,11 @@ long job, a second node changes nothing about it.
 
 ## Health checks and probes
 
-The check that ships with `Quartz.AspNetCore` asserts two things: that the scheduler is in a state that
-can fire, and that its job store answers a query. It reports *healthy* for a running scheduler whose
-store responds, *degraded* for one in standby, and *unhealthy* for one that was created but never
-started, is shutting down, has shut down, or whose store threw.
+The check that ships with `Quartz` asserts two things: that the scheduler is in a state that can fire,
+and that its job store answers a query. It reports *healthy* for a running scheduler whose store
+responds, *degraded* for one in standby, and *unhealthy* for one that was created but never started, is
+shutting down, has shut down, or whose store threw. It registers on the standard `IHealthChecksBuilder`
+and needs nothing from ASP.NET Core, so a worker on a `dotnet/runtime` image carries it too.
 
 <!-- snippet: sample_operations_readiness_probe -->
 ```csharp
