+++ src/Quartz.Tests.AspNetCore/Verify/PublicApiTest_Quartz.AspNetCore.verified.txt
@@ -3,20 +3,9 @@ namespace Quartz
 {
     public static class QuartzAspNetCoreConfigurationExtensions
     {
-        public static Microsoft.Extensions.DependencyInjection.IHealthChecksBuilder AddQuartz(this Microsoft.Extensions.DependencyInjection.IHealthChecksBuilder builder, System.Action<Quartz.QuartzHealthCheckOptions>? configure = null) { }
-        public static Microsoft.Extensions.DependencyInjection.IHealthChecksBuilder AddQuartz(this Microsoft.Extensions.DependencyInjection.IHealthChecksBuilder builder, string schedulerName, System.Action<Quartz.QuartzHealthCheckOptions>? configure = null) { }
-        public static Microsoft.Extensions.DependencyInjection.IServiceCollection AddQuartzHealthChecks(this Microsoft.Extensions.DependencyInjection.IServiceCollection services, System.Action<Quartz.QuartzHealthCheckOptions>? configure = null) { }
-        public static Quartz.IQuartzBuilder AddQuartzHealthChecks(this Quartz.IQuartzBuilder builder, System.Action<Quartz.QuartzHealthCheckOptions>? configure = null) { }
         public static Microsoft.Extensions.DependencyInjection.IServiceCollection AddQuartzHttpApi(this Microsoft.Extensions.DependencyInjection.IServiceCollection services, System.Action<Quartz.QuartzHttpApiOptions>? configure = null) { }
         public static Microsoft.AspNetCore.Builder.IEndpointConventionBuilder MapQuartzHttpApi(this Microsoft.AspNetCore.Routing.IEndpointRouteBuilder builder) { }
         public static Microsoft.AspNetCore.Builder.IEndpointConventionBuilder MapQuartzHttpApi(this Microsoft.AspNetCore.Routing.IEndpointRouteBuilder builder, string pattern) { }
-    }
-    public sealed class QuartzHealthCheckOptions
-    {
-        public QuartzHealthCheckOptions() { }
-        public Microsoft.Extensions.Diagnostics.HealthChecks.HealthStatus? FailureStatus { get; set; }
-        public string? Name { get; set; }
-        public System.Collections.Generic.List<string> Tags { get; }
     }
     public sealed class QuartzHttpApiOptions
     {
