+++ src/Quartz.Tests.Unit/HealthChecks/QuartzHealthCheckRegistrationTests.cs
@@ -1,10 +1,10 @@
+#nullable enable
 
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Diagnostics.HealthChecks;
 using Microsoft.Extensions.Options;
 
-
-namespace Quartz.Tests.AspNetCore.HealthChecks;
+namespace Quartz.Tests.Unit.HealthChecks;
 
 public class QuartzHealthCheckRegistrationTests
 {
