+++ src/Quartz.Tests.Unit/Verify/PublicApiTest_Quartz.verified.txt
@@ -1233,6 +1233,20 @@ namespace Quartz
             where T : Quartz.IJob { }
         public static Quartz.IQuartzBuilder UseSimpleTypeLoader(this Quartz.IQuartzBuilder builder) { }
     }
+    public static class QuartzHealthCheckExtensions
+    {
+        public static Microsoft.Extensions.DependencyInjection.IHealthChecksBuilder AddQuartz(this Microsoft.Extensions.DependencyInjection.IHealthChecksBuilder builder, System.Action<Quartz.QuartzHealthCheckOptions>? configure = null) { }
+        public static Microsoft.Extensions.DependencyInjection.IHealthChecksBuilder AddQuartz(this Microsoft.Extensions.DependencyInjection.IHealthChecksBuilder builder, string schedulerName, System.Action<Quartz.QuartzHealthCheckOptions>? configure = null) { }
+        public static Microsoft.Extensions.DependencyInjection.IServiceCollection AddQuartzHealthChecks(this Microsoft.Extensions.DependencyInjection.IServiceCollection services, System.Action<Quartz.QuartzHealthCheckOptions>? configure = null) { }
+        public static Quartz.IQuartzBuilder AddQuartzHealthChecks(this Quartz.IQuartzBuilder builder, System.Action<Quartz.QuartzHealthCheckOptions>? configure = null) { }
+    }
+    public sealed class QuartzHealthCheckOptions
+    {
+        public QuartzHealthCheckOptions() { }
+        public Microsoft.Extensions.Diagnostics.HealthChecks.HealthStatus? FailureStatus { get; set; }
+        public string? Name { get; set; }
+        public System.Collections.Generic.List<string> Tags { get; }
+    }
     public static class QuartzHostApplicationBuilderExtensions
     {
         public static Microsoft.Extensions.Hosting.IHostApplicationBuilder AddQuartz(this Microsoft.Extensions.Hosting.IHostApplicationBuilder builder, System.Action<Quartz.IQuartzBuilder>? configure = null) { }
