+++ src/Quartz/Diagnostics/HealthChecks/QuartzHealthCheck.cs
